dom/xhr/tests/test_sync_xhr_event_handling_switch_bcg.html

<!DOCTYPE HTML>
<!-- vim: set shiftwidth=2 tabstop=2 autoindent cindent expandtab: -->
<html>
<head>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css">
</head>
<body >
  <script type="text/javascript">
  SimpleTest.waitForExplicitFinish();

  let receivedMessage = [];
  const channel = new BroadcastChannel('taskManagerStatus');
  channel.onmessage = function(message) {
    receivedMessage.push(message.data);
  }

  let hasTestedSameBCG = false;
  function startSlowXHR(aWindowFeature) {
    var xhr = new XMLHttpRequest();
    window.open("file_sync_xhr_event_handling_switch_bcg_helper.html", "", aWindowFeature);
    xhr.open("GET", "slow.sjs", false);
    xhr.send(null);
    if (hasTestedSameBCG) {
      SimpleTest.isDeeply(receivedMessage, [true, false], "The first message should be true because the helper document was opened in the same BCG, and the second message should be false because it was opened in a different BCG");
      SimpleTest.finish();
    } else {
      hasTestedSameBCG = true;
      startSlowXHR("noopener");
    }
  }
  startSlowXHR("");
  </script>
</body>
</html>
